moves squidex entity to crn folder
