+++ packages/squidex/src/entities/crn/news.ts
@@ -1,4 +1,4 @@
-import { Rest, Entity, Graphql } from './common';
+import { Rest, Entity, Graphql } from '../common';
 
 export type NewsType = 'News' | 'Event' | 'Training';
