+++ packages/squidex/src/entities/index.ts
@@ -1,12 +1,2 @@
-export * from './calendar';
-export * from './lab';
-export * from './news';
-export * from './page';
-export * from './research-output';
-export * from './team';
-export * from './user';
-export * from './external-author';
-export * from './group';
-export * from './event';
-export * from './webhook';
-export * from './migration';
+export * from './common';
+export * from './crn';
